+++ src/law/cryptographic.rs
@@ -18,7 +18,7 @@ pub fn push_sq_to_hash(hash: u64, square: Option<&AbsoluteAddress2D>) -> u64 {
 pub fn pop_sq_from_hash(hash: u64) -> (u64, Option<AbsoluteAddress2D>) {
     // 0筋とか 0段とか 使ってないが、そのまま足す。
     // 0～100の101升と、ちょいなんで、128(=2^7) あれば十分
-    let adr = AbsoluteAddress2D::from_absolute_address((hash & 0b111_1111) as usize);
+    let adr = AbsoluteAddress2D::from_absolute_address((hash & 0b111_1111) as u8);
     (hash >> 7, adr)
 }
